Added some starter code for deep sleep, but doesn't work for C3. C3 does not have EXT0 wakeup

+++ Software/Projects/IR_Kill_SW_TX/src/main.cpp
@@ -28,7 +28,7 @@
 // Variables & Constants ------------------------------------------------------
 
 // IR control
-const uint32_t kill_code = 0xF; // Code to unalive the robot
+const uint8_t kill_code = 0xF; // Code to unalive the robot
 
 // Debounce
 bool lastButtonState = LOW; // Previous state of the button
